Verifier and public coin overhaul (#22)
In this PR I tried to address most other concerns I had about the state of the code. They mostly concern verifier.
* `FridaProof` made immutable and verification no longer consumes it.
* Verifier channel made internal to the verifier, similarly to what I did with prover.
* Removed traits `BaseFridaVerifier` and `BaseVerifierChannel`, their functionality merged into the base `impl`s

Another significant change is that fresh random coin is now predefined and is no longer passed by the caller. This is because both sides HAVE to agree on the initial state.

There are few test changes as well:
* `frida_verifier::tests` module flattened, previously it was wrapped in a `mod test { ... }` for no reason.
* Test type aliases and test-only functons moved to a dedicated `utils::test_utils` crate.